12.02 prtr

+++ src/dataset/coco.py
@@ -175,8 +175,8 @@ class COCODataset(JointsDataset):
             if max(obj["keypoints"]) == 0:
                 continue
 
-            joints_3d = np.zeros((self.num_joints, 3), dtype=np.float_)
-            joints_3d_vis = np.zeros((self.num_joints, 3), dtype=np.float_)
+            joints_3d = np.zeros((self.num_joints, 3), dtype=np.float32)
+            joints_3d_vis = np.zeros((self.num_joints, 3), dtype=np.float32)
             for ipt in range(self.num_joints):
                 joints_3d[ipt, 0] = obj["keypoints"][ipt * 3 + 0]
                 joints_3d[ipt, 1] = obj["keypoints"][ipt * 3 + 1]
@@ -262,8 +262,8 @@ class COCODataset(JointsDataset):
             num_boxes = num_boxes + 1
 
             center, scale = self._box2cs(box)
-            joints_3d = np.zeros((self.num_joints, 3), dtype=np.float_)
-            joints_3d_vis = np.ones((self.num_joints, 3), dtype=np.float_)
+            joints_3d = np.zeros((self.num_joints, 3), dtype=np.float32)
+            joints_3d_vis = np.ones((self.num_joints, 3), dtype=np.float32)
             kpt_db.append({
                 "image": img_name,
                 "center": center,
@@ -387,7 +387,7 @@ class COCODataset(JointsDataset):
             _key_points = np.array(
                 [img_kpts[k]["keypoints"] for k in range(len(img_kpts))])
             key_points = np.zeros((_key_points.shape[0], self.num_joints * 3),
-                                  dtype=np.float_)
+                                  dtype=np.float32)
 
             for ipt in range(self.num_joints):
                 key_points[:, ipt * 3 + 0] = _key_points[:, ipt, 0]
